RInside-package {RInside}	R Documentation

Embedding R in C++ applications

Description

The RInside package makes it easier to embed R in your C++ applications. There is no code you would execute directly from the R environment. Rather, you write C++ programs that embed R which is illustrated by some the included examples.

Author(s)

Dirk Eddelbuettel and Romain Francois

[Package RInside version 0.2.13 Index]